% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_bounds}
\alias{check_bounds}
\title{Check if parameters are up against the bounds}
\usage{
check_bounds(opt, lower, upper)
}
\arguments{
\item{opt}{an optimized TMB object.}

\item{lower}{a vector of lower bounds.}

\item{upper}{a vector of upper bounds.}
}
\value{
a \code{data.frame}.
}
\description{
A \code{data.frame} containing the parameters that are against their lower or
upper bound.
}
