% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{check_grid}
\alias{check_grid}
\title{Check estimability of the grid}
\usage{
check_grid(grid)
}
\arguments{
\item{grid}{a \code{list} of model runs.}
}
\value{
a \code{list} of \code{list}s.
}
\description{
Check estimability of the grid
}
