% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cmb}
\alias{cmb}
\title{Helper to make closure}
\usage{
cmb(f, d)
}
\arguments{
\item{f}{a \code{vector} of midpoints.}

\item{d}{natural mortality at the reference size.}
}
\value{
a \code{vector}.
}
\description{
Helper to make closure
}
