% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_oa_request.R
\name{doi_oa_request}
\alias{doi_oa_request}
\title{doi_oa_request}
\usage{
doi_oa_request(doi, ...)
}
\arguments{
\item{doi}{A DOI}

\item{...}{Additional arguments to be passed to \code{oa_request()} (e.g., verbose = \code{T}/\code{F})}
}
\value{
A list containing the API call results
}
\description{
Retrieve data for a work from the OpenAlex API based on a DOI.
}
\details{
(Details forthcoming.)
}
\examples{
# https://doi.org/10.4324/9780203873366 and doi:10.4324/9780203873366 work as well
doi <- '10.4324/9780203873366'
results <- doi_oa_request(doi)

}
\seealso{
Other OpenAlex: 
\code{\link{name_search_oa_request}()},
\code{\link{oa_request}()},
\code{\link{uva_oa_request}()}
}
\concept{OpenAlex}
