% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{evaluate_priors}
\alias{evaluate_priors}
\title{Evaluate priors}
\usage{
evaluate_priors(parameters, priors)
}
\arguments{
\item{parameters}{A \code{list} specifying the parameters to be passed to \code{MakeADFun}. Can be generated using the \code{get_parameters()} function.}

\item{priors}{A \code{list} of named \code{list}s specifying priors for the parameters. Can be generated using the \code{get_priors()} function.}
}
\value{
A \code{numeric} value.
}
\description{
Evaluate priors
}
\details{
Supported distributions are \code{normal}, \code{student} (three degrees of
freedom), \code{lognormal}, and \code{beta} (mean and precision). Priors act on the
stored parameter scale; for example, a normal prior on \code{log_B0} is a normal
density on log spawning output. Locations and scales must be scalars or
match the parameter block length. Indices identify blocks in \code{parameters}.
}
\examples{
{
  parameters <- list(log_B0 = log(1e6))
  priors <- list(
    log_B0 = list(
      type = "normal", par1 = log(1e6), par2 = 1,
      index = which("log_B0" == names(parameters))
    )
  )
  evaluate_priors(parameters, priors)
}

}
