% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_bounds}
\alias{get_bounds}
\title{Get default parameter bounds}
\usage{
get_bounds(obj, parameters)
}
\arguments{
\item{obj}{a \code{list} specifying the AD object created using the \code{MakeADFun} function.}

\item{parameters}{The parameter list used to construct \code{obj}. Retained
for API compatibility.}
}
\value{
a \code{data.frame} of parameter bounds.
}
\description{
Get \code{data.frame} of default parameter bounds.
}
