% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{get_dl}
\alias{get_dl}
\title{Obtain dl}
\usage{
get_dl(length_mu_ysa, length_sd_a)
}
\arguments{
\item{length_mu_ysa}{an \code{array} containing the mean length at age (a) for each year (y) and season (s).}

\item{length_sd_a}{a \code{vector} containing the standard deviation of the mean length at age.}
}
\value{
an \code{array}.
}
\description{
Obtain dl for use later in calculating phi. This function is used within
the \code{get_data()} function so it is generally not needed directly.
}
\details{
Obtain dl for use later in calculating phi(a, y). From the
PRELIMINARY_CALCS_SECTION of \code{opalmod.tpl}. More fine-scale than
lenage_dist_syal. Needs a bit more detail to get it right.
Integrates over length-at-age distribution to get phi(age,year).
}
\examples{
\dontrun{
  length_mu_ysa <- get_length_at_age(length_mean = opal::length_mean)
  dl_yal <- get_dl(length_mu_ysa = length_mu_ysa, length_sd_a = opal::length_sd$SD)
}

}
