% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{get_grid}
\alias{get_grid}
\title{Set up a grid}
\usage{
get_grid(
  parameters,
  m0 = c(0.5, 0.4, 0.45),
  m10 = c(0.065, 0.085, 0.105),
  h = c(0.55, 0.63, 0.72, 0.8),
  psi = c(1.75, 1.5, 2)
)
}
\arguments{
\item{parameters}{a \code{list} containing the parameter inputs.}

\item{m0}{the M0 values to be included in the grid.}

\item{m10}{the M10 values to be included in the grid.}

\item{h}{the h values to be included in the grid.}

\item{psi}{the psi values to be included in the grid.}
}
\value{
a \code{list} of parameter inputs ready to be passed to \code{MakeADFun}.
}
\description{
Set up a grid
}
\examples{
\dontrun{
  #parameters <- get_parameters(data = data)
  #grid_parameters <- get_grid(parameters = parameters)
}

}
