% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{get_length_at_age}
\alias{get_length_at_age}
\title{Obtain the length at age array}
\usage{
get_length_at_age(length_mean)
}
\arguments{
\item{length_mean}{a \code{data.frame} containing the mean length at age (a) for each year (y) and season (s).}
}
\value{
an \code{array}.
}
\description{
Obtain the mean length (cm) at age for each year and season. This function is
used within \code{get_data} so is generally not needed directly. It simply
structures the input \code{data.frame} into an \code{array}.
}
\details{
The input \code{data.frame} must contain the mean length (cm) at age arranged
into the columns \code{Year}, \code{Season}, and ages 0 to 30:\tabular{lllllll}{
   Year \tab Season \tab 0 \tab 1 \tab ... \tab 29 \tab 30 \cr
   1931 \tab 1 \tab 45 \tab 57.1 \tab ... \tab 74.1 \tab 88.8 \cr
   1932 \tab 1 \tab 45 \tab 57.1 \tab ... \tab 74.1 \tab 88.8 \cr
   ... \tab ... \tab ... \tab ... \tab ... \tab ... \tab ... \cr
   2021 \tab 2 \tab 50 \tab 68.9 \tab ... \tab 184.0 \tab 184.0 \cr
   2022 \tab 2 \tab 50 \tab 68.9 \tab ... \tab 184.0 \tab 184.0 \cr
}
}
\examples{
\dontrun{
  length_mu_ysa <- get_length_at_age(length_mean = opal::length_mean)
}

}
