% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looic.R
\name{get_loo}
\alias{get_loo}
\title{Obtain the leave-one-out information criterion (LOO IC)}
\usage{
get_loo(data, object, posterior)
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{object}{A \code{list} specifying the AD object created using \code{MakeADFun}.}

\item{posterior}{An \code{snutsfit} object created using the \code{sample_snuts} function.}
}
\value{
a \code{psis_loo} object.
}
\description{
This function extracts point-wise log-likelihood values and then calculates LOO statistics.
}
