% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-mortality.R
\name{get_M_length}
\alias{get_M_length}
\title{Natural mortality at length}
\usage{
get_M_length(min_age, max_age, age_increase_M, m0, m30, length_mu_ysa)
}
\arguments{
\item{min_age}{minimum model age.}

\item{max_age}{maximum model age.}

\item{age_increase_M}{age at which M begins to increase again.}

\item{m0}{M at age-1 and 2.}

\item{m30}{M at age-30 (terminal age M).}

\item{length_mu_ysa}{M at age-30 (terminal age M).}
}
\value{
vector of M at age.
}
\description{
Constructs a vector of M-at-age values using a declining early-age curve and late-age increase.
}
