% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_map}
\alias{get_map}
\title{Get default parameter mapping}
\usage{
get_map(parameters)
}
\arguments{
\item{parameters}{a \code{list} containing the initial parameter values to be passed to \code{MakeADFun}.}
}
\value{
a named \code{list} of parameter mapping.
}
\description{
Get a default parameter mapping. Parameter mapping is used by \code{MakeADFun}
to turn parameters on/off.
}
