% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{get_maturity_at_age}
\alias{get_maturity_at_age}
\title{Convert maturity-at-length to maturity-at-age}
\usage{
get_maturity_at_age(pla, maturity_at_length)
}
\arguments{
\item{pla}{Matrix (n_len x n_age). Probability of length at age (from
\code{\link{get_pla}}).}

\item{maturity_at_length}{Numeric vector (length n_len). Maturity at each
length bin (data).}
}
\value{
Numeric vector (length n_age). Maturity at each age.
}
\description{
Uses the probability-of-length-at-age matrix (PLA) to convert a
maturity-at-length vector to maturity-at-age:
\code{mat_a = t(pla) \%*\% mat_l}
}
