% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_mcmc_profile}
\alias{get_mcmc_profile}
\title{Run the dynamics}
\usage{
get_mcmc_profile(obj, mcmc, name)
}
\arguments{
\item{obj}{last year of model before projection}

\item{mcmc}{number of projection years}

\item{name}{number of projection years}
}
\value{
a \code{list} of derived values.
}
\description{
Run the dynamics
}
