% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Obtain samples from the posterior distribution of reported quantities}
\usage{
get_posterior(
  object,
  posterior,
  pars = "par_B0",
  iters = NULL,
  option = 2,
  type = "df"
)
}
\arguments{
\item{object}{A \code{list} specifying the AD object created using \code{MakeADFun}.}

\item{posterior}{An \code{rstan} objected created using the \code{tmbstan} function.}

\item{pars}{The parameter(s) to be extracted.}

\item{iters}{The number of iterations to be extracted.}

\item{option}{The parallel option to use.}

\item{type}{default data frame "df" or "array"}
}
\value{
A \code{data.frame} or \code{array}
}
\description{
When samples from the posterior distribution are obtained using \code{tmbstan},
only the model parameters can be accessed directly from the \code{stanfit}
object (i.e., derived quantities such as \code{par_B0} or \code{M_a} cannot be accessed
directly). Accessing derived quantities must be done using the this function.
}
