% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{get_recruitment}
\alias{get_recruitment}
\title{Calculate recruitment}
\usage{
get_recruitment(sbio, rdev, B0, alpha, beta, sigma_r = 0.6, bias_adj = 1)
}
\arguments{
\item{sbio}{Spawning biomass.}

\item{rdev}{Recruitment deviations.}

\item{B0}{Unfished biomass.}

\item{alpha, beta}{Beverton-Holt stock recruitment parameters.}

\item{sigma_r}{Lognormal SD of recruitment deviations.}

\item{bias_adj}{Bias adjustment scalar (typically year-specific) applied to
the lognormal correction term.}
}
\value{
Recruitment value (numeric).
}
\description{
Computes recruitment based on Beverton-Holt with depensation and log-normal deviations.
}
