% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{get_recruitment_prior}
\alias{get_recruitment_prior}
\title{Recruitment prior}
\usage{
get_recruitment_prior(rdev, sigma_r)
}
\arguments{
\item{rdev}{a \code{vector} of recruitment deviations.}

\item{sigma_r}{recruitment standard deviation.}
}
\value{
negative log-prior (scalar).
}
\description{
Calculates the prior for the recruitment deviations.
}
