% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{get_rho}
\alias{get_rho}
\title{Estimate temporal autocorrelation in recruitment deviations}
\usage{
get_rho(first_yr = 1931, last_yr = 2022, rdev)
}
\arguments{
\item{first_yr}{First model year.}

\item{last_yr}{Last model year.}

\item{rdev}{Vector of recruitment deviations.}
}
\value{
Estimated autocorrelation.
}
\description{
Calculates the AR1 autocorrelation coefficient (phi) for recruitment deviations.
}
\examples{
\dontrun{
  first_yr <- 1931
  last_yr <- 2022
  N <- length(first_yr:last_yr)
  rdev <- arima.sim(list(order = c(1, 0, 0), ar = 0.5), n = N)
  get_rho(first_yr, last_yr, rdev)
}

}
