% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{get_sd_at_age}
\alias{get_sd_at_age}
\title{Compute SD of length-at-age from CV1 and CV2}
\usage{
get_sd_at_age(mu_a, L1, L2, log_CV1, log_CV2)
}
\arguments{
\item{mu_a}{Numeric vector. Mean length at age (from \code{\link{get_growth}}).
May be AD.}

\item{L1}{Numeric. Length at age A1 (may be AD).}

\item{L2}{Numeric. Length at age A2 (may be AD).}

\item{log_CV1}{Numeric. CV at age A1 (may be AD).}

\item{log_CV2}{Numeric. CV at age A2 (may be AD).}
}
\value{
Numeric vector of SD at each age.
}
\description{
Linearly interpolates CV as a function of mean length, matching
SS3's \code{CV_Growth_Pattern = 2}.
}
