% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{get_weight_at_age}
\alias{get_weight_at_age}
\title{Obtain the mean weight at age}
\usage{
get_weight_at_age(length_mu_ysa, length_sd_a)
}
\arguments{
\item{length_mu_ysa}{an \code{array} containing the mean length at age (a) for each year (y) and season (s). This \code{array} can be generated using the function \code{get_length_at_age}.}

\item{length_sd_a}{a \code{vector} containing the standard deviation of mean length at age.}
}
\value{
an \code{array}.
}
\description{
Obtain the mean weight (kg) at age for each fishery each year. This function
is used within \code{get_data} so is generally not needed directly.
}
\examples{
length_mu_ysa <- get_length_at_age(length_mean = sbt::length_mean)
weight_fya <- get_weight_at_age(length_mu_ysa = length_mu_ysa, 
                                length_sd_a = sbt::length_sd$SD)

}
