% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{get_weight_at_length}
\alias{get_weight_at_length}
\title{Compute weight at each length bin midpoint}
\usage{
get_weight_at_length(len_mid, lw_a, lw_b)
}
\arguments{
\item{len_mid}{Numeric vector. Length bin midpoints (data).}

\item{lw_a}{Numeric. L-W scalar (data).}

\item{lw_b}{Numeric. L-W exponent (data).}
}
\value{
Numeric vector of weight at each length bin.
}
\description{
Compute weight at each length bin midpoint
}
