% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_to_snutsfit}
\alias{grid_to_snutsfit}
\title{Convert grid to an snutsfit object}
\usage{
grid_to_snutsfit(data, parameters, grid, grid_parameters, grid_cells)
}
\arguments{
\item{data}{a \code{list} of model runs.}

\item{parameters}{a \code{list} of model runs.}

\item{grid}{a \code{list} of model runs.}

\item{grid_parameters}{a \code{list} of model runs.}

\item{grid_cells}{a \code{list} of model runs.}
}
\value{
a \code{vector} of integers.
}
\description{
Convert grid to an snutsfit object
}
