% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_oa_request.R
\name{name_search_oa_request}
\alias{name_search_oa_request}
\title{name_search_oa_request}
\usage{
name_search_oa_request(name)
}
\arguments{
\item{name}{A search string (e.g., "sagan"; "hadley wickham")}
}
\value{
A list containing the API call results; each element of the outermost list contains data for the works associated with one matched author (some searchers return >1 author)
}
\description{
Retrieve data from the OpenAlex API for works with authors matching a search string.
}
\details{
(Details forthcoming.)
}
\examples{
results <- name_search_oa_request('carl sagan')

}
\seealso{
Other OpenAlex: 
\code{\link{doi_oa_request}()},
\code{\link{oa_request}()},
\code{\link{uva_oa_request}()}
}
\concept{OpenAlex}
