#-------------------------------------------------------------------------------
# Copyright (c) 2011 OBiBa. All rights reserved.
#  
# This program and the accompanying materials
# are made available under the terms of the GNU Public License v3.0.
#  
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.
#-------------------------------------------------------------------------------
\name{opal-package}
\alias{opal-package}
\alias{opal}
\docType{package}
\title{Opal Client for du rapporR}
\description{
Allows invoking Opal Web Services from directly from R. Also includes a DataShield client.
}
\author{
OBiBa <info@obiba.org>
}
