% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opal_baseline_parameters}
\alias{opal_baseline_parameters}
\title{Opal Baseline Parameters}
\format{
An object of class \code{list} of length 16.
}
\source{
\code{dev/vignettes/baseline.qmd}
}
\usage{
opal_baseline_parameters
}
\description{
The full parameter list used for testing in the opal baseline model.
}
\keyword{datasets}
