% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal-fit.R
\name{opal_fit_compatibility}
\alias{opal_fit_compatibility}
\title{Check compatibility of a saved opal fit}
\usage{
opal_fit_compatibility(x)
}
\arguments{
\item{x}{An \code{opal_fit} object.}
}
\value{
A list containing \code{compatible}, \code{errors}, and \code{warnings}.
}
\description{
Check compatibility of a saved opal fit
}
