% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal-fit.R
\name{opal_fit_report}
\alias{opal_fit_report}
\title{Recreate the fitted model report}
\usage{
opal_fit_report(x)
}
\arguments{
\item{x}{An \code{opal_fit} object.}
}
\value{
The named report list returned by \code{opal_model}.
}
\description{
Recreate the fitted model report
}
