% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{opal_globals}
\alias{opal_globals}
\title{The opal globals}
\usage{
opal_globals()
}
\value{
a \code{list} of functions to be passed to \code{sample_sparse_tmb} when doing MCMC.
}
\description{
The opal globals
}
