% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{opal_model}
\alias{opal_model}
\title{The opal model}
\usage{
opal_model(parameters, data)
}
\arguments{
\item{parameters}{a \code{list} of parameter values.}

\item{data}{a \code{list} of data inputs.}
}
\value{
the negative log-likelihood (NLL) value.
}
\description{
Obtain the negative log-likelihood (NLL) value from the opal model.
}
