% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_biomass_spawning}
\alias{plot_biomass_spawning}
\title{Plot spawning biomass}
\usage{
plot_biomass_spawning(
  data_list,
  object_list = NULL,
  relative = TRUE,
  labels = NULL
)
}
\arguments{
\item{data_list}{An Opal object, list of Opal objects, or list of model data lists.}

\item{object_list}{A list of AD objects created using \code{MakeADFun}.}

\item{relative}{Logical; plot spawning biomass relative to unfished biomass.}

\item{labels}{Optional labels for the model runs.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot spawning biomass or relative spawning biomass by year for one or more
model runs.
}
