% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{plot_dist_length_at_age}
\alias{plot_dist_length_at_age}
\title{Plot the distribution of length at age}
\usage{
plot_dist_length_at_age(data, years = c(1931, 2022))
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{years}{the years to show on the plot.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the distribution of length (cm) at age by year and season.
}
