% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial_numbers}
\alias{plot_initial_numbers}
\title{Plot initial numbers}
\usage{
plot_initial_numbers(data, object = NULL)
}
\arguments{
\item{data}{An \code{opal_obj}, legacy \code{opal_fit}, or model data list.}

\item{object}{Optional for Opal objects. The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot initial population numbers by age.
}
