% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{plot_length_at_age}
\alias{plot_length_at_age}
\title{Plot length at age}
\usage{
plot_length_at_age(data, years = c(1931, 2022))
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{years}{the years to show on the plot.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the mean length (cm) at age by year and season. Plus and minus one
standard deviation is also plotted.
}
