% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lf}
\alias{plot_lf}
\title{Plot LF}
\usage{
plot_lf(
  data,
  object,
  posterior = NULL,
  probs = c(0.025, 0.975),
  years = NULL,
  fishery = "LL1",
  ...
)
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{object}{a \code{list} specifying the AD object created using \code{MakeADFun}.}

\item{posterior}{an \code{rstan} objected created using the \code{tmbstan} function.}

\item{probs}{a numeric vector of probabilities with values in \code{[0,1]} for plotting quantiles of the posterior distribution.}

\item{years}{the years to plot.}

\item{fishery}{the fishery to plot (LL1, LL2, LL3, or LL4).}

\item{...}{additional parameters passed on to \code{facet_wrap}.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot LF fit.
}
