% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looic.R
\name{plot_loo}
\alias{plot_loo}
\title{Plot PSIS LOO}
\usage{
plot_loo(x, exclude = NULL)
}
\arguments{
\item{x}{A \code{psis_loo} object.}

\item{exclude}{Any variables to exclude from the plot (e.g., aerial, af, lf, cpue).}
}
\value{
a \code{ggplot2}.
}
\description{
This is the same as \code{plot(loo1)} except it colours/groups the different
types of data.
}
