% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-mortality.R
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality}
\usage{
plot_natural_mortality(data, object)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{object}{The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot natural mortality (M) by age.
}
