% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal-residuals.R
\name{plot_osa_residuals}
\alias{plot_osa_residuals}
\title{Plot stored OSA residuals}
\usage{
plot_osa_residuals(x, name = "osa", type = c("time", "qq"))
}
\arguments{
\item{x}{An Opal object containing stored OSA diagnostics.}

\item{name}{Name of the stored diagnostic result.}

\item{type}{Plot residuals against model year or as normal quantiles.}
}
\value{
A ggplot, faceted by dataset.
}
\description{
Plot stored OSA residuals
}
\seealso{
Other assessment diagnostics:
\code{\link[=opal_diagnose]{opal_diagnose()}},
\code{\link[=opal_osa]{opal_osa()}},
\code{\link[=plot_composition]{plot_composition()}},
\code{\link[=plot_osa_sdnr]{plot_osa_sdnr()}},
\code{\link[=plot_prior_posterior]{plot_prior_posterior()}}
}
\concept{assessment diagnostics}
