% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Run the dynamics}
\usage{
plot_profile(x, y = NULL, xlab = NULL, rescale = TRUE)
}
\arguments{
\item{x}{MLE profile}

\item{y}{MCMC profile}

\item{xlab}{number of projection years}

\item{rescale}{number of projection years}
}
\value{
a \code{list} of derived values.
}
\description{
Run the dynamics
}
