% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(
  data,
  object,
  posterior = NULL,
  probs = c(0.025, 0.975),
  years = 2013:2022,
  fisheries = c("LL1", "LL2", "LL3", "LL4", "Indonesian", "Australian", "CPUE"),
  ...
)
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{object}{a \code{list} specifying the AD object created using \code{MakeADFun}.}

\item{posterior}{an \code{rstan} objected created using the \code{tmbstan} function.}

\item{probs}{a numeric vector of probabilities with values in \code{[0, 1]} for plotting quantiles of the posterior distribution.}

\item{years}{the years to show on the plot.}

\item{fisheries}{the fisheries to show on the plot.}

\item{...}{options passed on to \code{geom_density_ridges}.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot selectivity by fishery, year, and age.
}
