% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-weight.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Plot weight at age}
\usage{
plot_weight_at_age(data, years = c(1931, 2000, 2010, 2022))
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{years}{the years to show on the plot.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the mean weight (kg) at age by fishery for a subset of years.
}
