% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_selectivity}
\alias{project_selectivity}
\title{Project selectivity}
\usage{
project_selectivity(
  data,
  obj = NULL,
  first_yr = NULL,
  last_yr = NULL,
  n_proj = 5,
  n_iter = 1
)
}
\arguments{
\item{data}{An Opal object or model data list.}

\item{obj}{a \code{list} of parameter values.}

\item{first_yr}{the first year sampled. Defaults to the first model year.}

\item{last_yr}{the last year.}

\item{n_proj}{the number of projection years.}

\item{n_iter}{the number of simulated trajectories.}
}
\value{
An array of projected selectivity by iteration, fishery, year, and
age.
}
\description{
Project selectivity
}
