% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{rerun_grid}
\alias{rerun_grid}
\title{Run a grid again}
\usage{
rerun_grid(grid, bounds, control = list(eval.max = 10000, iter.max = 10000))
}
\arguments{
\item{grid}{a \code{list} of parameter inputs for each grid cell. This can be created using \code{get_grid}.}

\item{bounds}{the lower and upper parameter bounds created using the \code{get_bounds} function.}

\item{control}{a \code{list} defining the controls passed to \code{nlminb}.}
}
\value{
a \code{list} of grid cells.
}
\description{
Fit a model for each grid cell defined in \code{grid} where each grid cell has a
different combination of fixed parameter values.
}
