% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{run_grid}
\alias{run_grid}
\title{Run a grid}
\usage{
run_grid(
  data,
  grid_parameters,
  bounds,
  map = list(),
  random = c(),
  control = list(eval.max = 10000, iter.max = 10000)
)
}
\arguments{
\item{data}{a \code{list} containing the data created using the \code{get_data} function.}

\item{grid_parameters}{a \code{list} of parameter inputs for each grid cell. This can be created using \code{get_grid}.}

\item{bounds}{the lower and upper parameter bounds created using the \code{get_bounds} function.}

\item{map}{a \code{list} defining how to optionally collect and fix parameters.}

\item{random}{a character \code{vector} defining the random effect parameters.}

\item{control}{a \code{list} defining the controls passed to \code{nlminb}.}
}
\value{
a \code{list} of grid cells.
}
\description{
Fit a model for each grid cell defined in \code{grid} where each grid cell has a
different combination of fixed parameter values.
}
