% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{sample_grid}
\alias{sample_grid}
\title{Sample from the grid}
\usage{
sample_grid(grid, n_samples = 2000)
}
\arguments{
\item{grid}{a \code{list} of model runs.}

\item{n_samples}{the number of samples from the grid.}
}
\value{
a \code{vector} of integers.
}
\description{
Sample from the grid
}
