% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{sbtprofile}
\alias{sbtprofile}
\title{Adaptive likelihood profiling}
\usage{
sbtprofile(
  obj,
  name,
  lincomb,
  h = 1e-04,
  ytol = 2,
  ystep = 0.1,
  maxit = ceiling(5 * ytol/ystep),
  parm.range = c(-Inf, Inf),
  slice = FALSE,
  adaptive = TRUE,
  trace = TRUE
)
}
\arguments{
\item{obj}{Object from MakeADFun that has been optimized.}

\item{name}{Name or index of a parameter to profile.}

\item{lincomb}{Optional linear combination of parameters to profile. By default a unit vector corresponding to name.}

\item{h}{Initial adaptive stepsize on parameter axis.}

\item{ytol}{Adjusts the range of the likelihood values.}

\item{ystep}{Adjusts the resolution of the likelihood profile.}

\item{maxit}{Max number of iterations for adaptive algorithm.}

\item{parm.range}{Valid parameter range.}

\item{slice}{Do slicing rather than profiling?}

\item{adaptive}{Logical; Use adaptive step size?}

\item{trace}{Trace progress? (TRUE, or a numeric value of 1, gives basic tracing: numeric values > 1 give more information).}
}
\value{
a \code{vector} penalty
}
\description{
Calculate 1D likelihood profiles with respect to single parameters or more
generally, with respect to arbitrary linear combinations of parameters (e.g. contrasts).
}
