% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{sel_length}
\alias{sel_length}
\title{Selectivity-at-length for a single fishery, dispatched on type code}
\usage{
sel_length(len, par, sel_type)
}
\arguments{
\item{len}{Numeric vector of length-bin midpoints.}

\item{par}{Numeric vector of length 6 containing selectivity parameters.}

\item{sel_type}{Selectivity type: 1 = logistic, 2 = double-normal, 3 =
double Richards.}
}
\value{
Numeric vector of selectivity values.
}
\description{
Selectivity-at-length for a single fishery, dispatched on type code
}
