% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_uva_institutions}
\alias{short_uva_institutions}
\title{short_uva_institutions}
\format{
A data frame
\describe{
\item{\code{id}}{OpenAlex ID}
\item{\code{institution}}{Institution name}
}
}
\source{
\url{https://github.com/jacob-gg}
}
\usage{
short_uva_institutions
}
\description{
Smaller version of uva_institutions; used for package testing.
}
\keyword{datasets}
