% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uva_institutions}
\alias{uva_institutions}
\title{uva_institutions}
\format{
A data frame
\describe{
\item{\code{id}}{OpenAlex ID}
\item{\code{institution}}{Institution name}
}
}
\source{
\url{https://github.com/jacob-gg}
}
\usage{
uva_institutions
}
\description{
OpenAlex IDs and names of UVA-affiliated institutions.
}
\keyword{datasets}
