% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uva_oa_request.R
\name{uva_oa_request}
\alias{uva_oa_request}
\title{uva_oa_request}
\usage{
uva_oa_request(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{oa_request()} (e.g., verbose = \code{T}/\code{F})}
}
\value{
A list containing the data from the OpenAlex API on all works affiliated with UVA institutions
}
\description{
Retrieve data for works affiliated with UVA institutions from the OpenAlex API.
}
\details{
(Details forthcoming.)
}
\examples{
\dontrun{uva_oa_request()}

}
\seealso{
Other OpenAlex: 
\code{\link{doi_oa_request}()},
\code{\link{oa_request}()}
}
\concept{OpenAlex}
