% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_oa_request.R
\name{doi_oa_request}
\alias{doi_oa_request}
\title{doi_oa_request}
\usage{
doi_oa_request(doi)
}
\arguments{
\item{doi}{A DOI}
}
\value{
A list containing the API call results
}
\description{
Retrieve data from the OpenAlex API based on a DOI
}
\details{
(Details forthcoming.)
}
