% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{evaluate_priors}
\alias{evaluate_priors}
\title{Evaluate priors}
\usage{
evaluate_priors(parameters, priors)
}
\arguments{
\item{parameters}{A \code{list} specifying the parameters to be passed to \code{MakeADFun}. Can be generated using the \code{get_parameters()} function.}

\item{priors}{A \code{list} of named \code{list}s specifying priors for the parameters. Can be generated using the \code{get_priors()} function.}
}
\value{
A \code{numeric} value.
}
\description{
Evaluate priors
}
\examples{
\dontrun{
  parameters <- list(par_log_m4 = log(0.167))
  priors <- list(
    par_log_m4 = list(type = "normal", par1 = log(0.12), par2 = 0.4, 
                      index = which("par_log_m4" == names(parameters)))
  )
  evaluate_priors(parameters, priors)
}

}
