% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{get_cpue_like}
\alias{get_cpue_like}
\title{CPUE index likelihood}
\usage{
get_cpue_like(
  data,
  parameters,
  number_ysa,
  sel_fya,
  weight_fya,
  creep_init = 1
)
}
\arguments{
\item{data}{a \code{list} of data inputs (cpue_data, cpue_switch, etc.).}

\item{parameters}{a \code{list} of parameter values (log_cpue_tau, log_cpue_omega, cpue_creep, log_cpue_q, etc.).}

\item{number_ysa}{a 3D \code{array} \verb{[n_year, n_season, n_age]} of numbers-at-age.}

\item{sel_fya}{a 3D \code{array} \verb{[n_fishery, n_year, n_age]} of selectivity by fishery, year, and age.}

\item{weight_fya}{a 3D \code{array} \verb{[n_fishery, n_year, n_age]} of weight-at-age by fishery and year.}

\item{creep_init}{scalar initialization value for creeping adjustment (default 1).}
}
\value{
a \code{numeric} vector of negative log-likelihood contributions.
}
\description{
Computes the likelihood for a standardized CPUE index using a log-linear model.
}
