% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{get_cpue_like}
\alias{get_cpue_like}
\title{CPUE index likelihood}
\usage{
get_cpue_like(data, parameters, number_ysa, sel_fya, creep_init = 1)
}
\arguments{
\item{data}{Integer switch to activate the likelihood.}

\item{parameters}{a \code{vector} of year indices for CPUE observations.}

\item{number_ysa}{a 3D \code{array} year, season, age of numbers-at-age.}

\item{sel_fya}{a 3D \code{array} of selectivity by fishery, year, and age.}
}
\value{
a \code{list} with predicted CPUE, residuals, and likelihood vector.
}
\description{
Computes the likelihood for a standardized CPUE index using a log-linear model.
}
