% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Set up the data input file}
\usage{
get_data(data_in)
}
\arguments{
\item{data_in}{a \code{list} containing the data inputs.}
}
\value{
a \code{list} ready to be passed to \code{MakeADFun}.
}
\description{
Set up the data input file to be passed to \code{MakeADFun}. This function
runs data cross validation tests and appends several inputs to the data list
including model dimensions and processed inputs:
}
\details{
\itemize{
\item \code{n_year}: dervied from \code{first_yr} and \code{last_yr}
\item \code{n_season}: set to 2
\item \code{n_length}: not in use
\item \code{n_age}: derived from \code{min_age} and \code{max_age}
\item \code{n_fishery}: set to 6
\item \code{age_a}: sequence of modeled ages derived from \code{min_age} and \code{max_age}
\item \code{length_mu_ysa}: derived from the \code{length_mean} input
\item \code{length_sd_a}: derived from the \code{length_sd} input
\item \code{dl_yal}: derived from \code{length_mu_ysa} and \code{length_sd_a}
\item \code{weight_fya}: derived from \code{length_mu_ysa} and \code{length_sd_a}
\item \code{catch_obs_ysf}: derived from \code{catch}, \code{catch_UA}, \code{scenarios_LL1}, and \code{scenarios_surf}
\item \code{sel_change_year_fy}: derived from \code{sel_change_sd_fy}
}

This function produces the data input file to be passed to \code{MakeADFun}.
}
