% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{get_initial_numbers}
\alias{get_initial_numbers}
\title{Initial numbers and Beverton-Holt parameters}
\usage{
get_initial_numbers(
  B0,
  h,
  M_a,
  spawning_potential_a,
  init_F_f = NULL,
  sel_fa = NULL
)
}
\arguments{
\item{B0}{Unfished spawning biomass.}

\item{h}{Beverton-Holt steepness parameter.}

\item{M_a}{a \code{vector} of natural mortality at age.}

\item{spawning_potential_a}{a \code{vector} of spawning potential at age
(maturity × fecundity).}

\item{init_F_f}{an optional \code{vector} of initial fishing mortality by
fishery.}

\item{sel_fa}{an optional matrix of selectivity-at-age with dimensions
\code{[n_fishery, n_age]}.}
}
\value{
A list containing:
\describe{
\item{Ninit}{Initial numbers-at-age (vector).}
\item{R0}{Unfished recruitment (scalar).}
\item{alpha}{BH alpha parameter.}
\item{beta}{BH beta parameter.}
}
}
\description{
Computes the initial equilibrium numbers-at-age, unfished recruitment (R0),
and Beverton-Holt stock-recruitment parameters.
}
