% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{get_length_like}
\alias{get_length_like}
\title{Length Composition Likelihood}
\usage{
get_length_like(data, parameters, log_lf_alpha, catch_pred_fya, alk_ysal)
}
\arguments{
\item{catch_pred_fya}{3D array of predicted catch.}

\item{alk_ysal}{4D array year, season, age, length_bin of ALKs.}

\item{lf_switch}{Vectors of indices for observations.}

\item{removal_switch_f}{Vectors of indices for observations.}

\item{lf_year, lf_season, lf_fishery}{Vectors of indices for observations.}

\item{lf_minbin}{Minimum size bin to be aggregated.}

\item{lf_obs}{Matrix of observed length proportions.}

\item{lf_n}{Vector of effective sample sizes.}

\item{par_log_lf_alpha}{Vector of effective sample sizes.}
}
\value{
List with predicted compositions and likelihood contributions.
}
\description{
Computes likelihood for observed length compositions using ALKs.
}
