% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get default initial parameter values}
\usage{
get_parameters(data)
}
\arguments{
\item{data}{a \code{list} containing the data inputs.}
}
\value{
a \code{list} of initial parameter values.
}
\description{
Get \code{list} of default initial parameter values.
}
