% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_priors}
\alias{get_priors}
\title{Get priors}
\usage{
get_priors(parameters, data = NULL)
}
\arguments{
\item{parameters}{A \code{list} specifying the parameters to be passed to \code{MakeADFun}. Can be generated using the \code{get_parameters()} function.}

\item{data}{A \code{list} of data inputs (optional). Used to retrieve prior
center values for growth/variability parameters (e.g.,
\code{data$prior_log_L1_mean}).}
}
\value{
A \code{list} of priors.
}
\description{
Don't include priors for recruitment deviates (par_rdev_y) or selectivity
(e.g., par_log_sel_1) here because they are dealt with in the \code{get_recruitment_prior()}
and \code{get_selectivity_prior()} functions.
}
