% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_request.R
\name{oa_request}
\alias{oa_request}
\title{oa_request}
\usage{
oa_request(
  query,
  use_fast_api_pool = T,
  remove_duplicates = T,
  verbose = T,
  return_one_pg = F
)
}
\arguments{
\item{query}{A URL (string) beginning with "http(s)://api.openalex.org"}

\item{use_fast_api_pool}{Logical value indicating whether to call to the OpenAlex API polite pool (faster) or common pool (slower)}

\item{remove_duplicates}{Logical value indicating whether to check for and remove duplicate records at the end of the call}

\item{verbose}{Logical value indicating whether to print API call details to the console}

\item{return_one_pg}{Logical value indicating whether to only return one page of results from the API (for testing and debugging purposes)}
}
\value{
A list containing the API call results
}
\description{
Retrieve data from the OpenAlex API using a custom query.
}
\details{
(Additional details forthcoming.)
}
\examples{
query <- 'https://api.openalex.org/authors/A5069290754'
sagan_dat <- oa_request(query)

}
\seealso{
Other OpenAlex: 
\code{\link{doi_oa_request}()},
\code{\link{uva_oa_request}()}
}
\concept{OpenAlex}
