% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_request.R
\name{oa_request}
\alias{oa_request}
\title{oa_request}
\usage{
oa_request(query)
}
\arguments{
\item{query}{A single URL string beginning with "http(s)://api.openalex.org"}
}
\value{
A list containing the API call results
}
\description{
Retrieve data from the OpenAlex API.
}
\details{
(Details forthcoming.)
}
\examples{
# query <- 'https://api.openalex.org/authors/A3184395717'
# sagan_dat <- oa_request(query)

}
