% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opal-fit.R
\name{opal_as_tmbfit}
\alias{opal_as_tmbfit}
\title{Convert normalized opal posterior draws to a tmbfit}
\usage{
opal_as_tmbfit(x)
}
\arguments{
\item{x}{An \code{opal_fit} with MCMC output or an \code{opal_mcmc} object.}
}
\value{
A list inheriting from \code{tmbfit}.
}
\description{
Convert normalized opal posterior draws to a tmbfit
}
