% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opal_baseline_data}
\alias{opal_baseline_data}
\title{Opal Baseline Data}
\format{
An object of class \code{list} of length 68.
}
\source{
\code{vignettes/baseline.Rmd}
}
\usage{
opal_baseline_data
}
\description{
The full data list used for testing in the opal baseline model.
}
\keyword{datasets}
