% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opal_baseline_map}
\alias{opal_baseline_map}
\title{Opal Baseline Map}
\format{
An object of class \code{list} of length 13.
}
\source{
\code{vignettes/baseline.qmd}
}
\usage{
opal_baseline_map
}
\description{
The parameter map used for testing in the opal baseline model.
}
\keyword{datasets}
