% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_biomass_spawning}
\alias{plot_biomass_spawning}
\title{Plot spawning biomass}
\usage{
plot_biomass_spawning(
  data_list,
  object_list,
  posterior = NULL,
  probs = c(0.025, 0.975),
  relative = TRUE,
  labels = NULL
)
}
\arguments{
\item{data_list}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{object_list}{a \code{list} specifying the AD object created using the \code{MakeADFun} function.}

\item{posterior}{an \code{rstan} objected created using the \code{tmbstan} function.}

\item{probs}{a numeric vector of probabilities with values in \code{[0,1]} for plotting quantiles of the posterior distribution. Defaults to the 90\% credible interval.}

\item{relative}{if the plot should be relative spawning biomass.}

\item{labels}{a \code{vector} of labels for the model runs.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot the spawning biomass (tonnes) or relative spawning biomass by year for a
single model run, a grid of model runs, or an MCMC.
}
