% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catch}
\usage{
plot_catch(data, obj = NULL, plot_resid = FALSE)
}
\arguments{
\item{data}{An \code{opal_obj}, legacy \code{opal_fit}, or model data list.}

\item{obj}{Optional for Opal objects. The AD object created using \code{MakeADFun}.}

\item{plot_resid}{Logical; plot catch residuals instead of observed and
predicted catch.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot catch by year and fishery.
}
