% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catch}
\usage{
plot_catch(data, obj, plot_resid = FALSE)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{obj}{The AD object created using \code{MakeADFun}.}

\item{plot_resid}{Logical; plot catch residuals instead of observed and
predicted catch.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot catch by year and fishery.
}
