% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot CPUE}
\usage{
plot_cpue(data, object = NULL)
}
\arguments{
\item{data}{An \code{opal_obj}, legacy \code{opal_fit}, or model data list.}

\item{object}{Optional for Opal objects. The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot observed and predicted CPUE by season and fishery.
}
\examples{
assessment <- opal_read(system.file("extdata", "opaka_quickstart_fit.rds",
                                   package = "opal"))
plot_cpue(assessment)
}
