% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot CPUE}
\usage{
plot_cpue(data, object)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{object}{The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot observed and predicted CPUE by season and fishery.
}
