% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hrate}
\alias{plot_hrate}
\title{Plot harvest rate}
\usage{
plot_hrate(data, object, years = NULL, ...)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{object}{The AD object created using \code{MakeADFun}.}

\item{years}{Optional years to show. The default plots every model year from
the first catch year onwards.}

\item{...}{Options passed to \code{geom_density_ridges}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot harvest rate by year, season, and age.
}
