% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial_numbers}
\alias{plot_initial_numbers}
\title{Plot initial numbers}
\usage{
plot_initial_numbers(data, object)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{object}{The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot initial population numbers by age.
}
