% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial_numbers}
\alias{plot_initial_numbers}
\title{Plot initial numbers}
\usage{
plot_initial_numbers(data, object, posterior = NULL, probs = c(0.025, 0.975))
}
\arguments{
\item{data}{a \code{list} containing the data that was passed to \code{MakeADFun}.}

\item{object}{a \code{list} specifying the AD object created using \code{MakeADFun}.}

\item{posterior}{an \code{rstan} objected created using the \code{tmbstan} function.}

\item{probs}{a numeric vector of probabilities with values in \code{[0,1]} for plotting quantiles of the posterior distribution.}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot initial numbers by age.
}
