% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment}
\usage{
plot_recruitment(data, object)
}
\arguments{
\item{data}{A model data list passed to \code{MakeADFun}.}

\item{object}{The AD object created using \code{MakeADFun}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot recruitment by year.
}
