% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_dynamics}
\alias{project_dynamics}
\title{Project dynamics}
\usage{
project_dynamics(
  data,
  object,
  mcmc = NULL,
  n_proj = 5,
  n_iter = 1,
  rdev_y,
  sel_fya,
  catch_ysf
)
}
\arguments{
\item{data}{a \code{list} of parameter values.}

\item{object}{a \code{list} of parameter values.}

\item{mcmc}{a \code{list} of parameter values.}

\item{n_proj}{the number of projection years.}

\item{n_iter}{a \code{list} of inputs.}

\item{rdev_y}{a \code{list} of inputs.}

\item{sel_fya}{a \code{list} of inputs.}

\item{catch_ysf}{a \code{list} of inputs.}
}
\value{
the negative log-likelihood (NLL) value.
}
\description{
Project dynamics
}
