% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_rec_devs}
\alias{project_rec_devs}
\title{Project recruitment deviates}
\usage{
project_rec_devs(
  data,
  obj,
  mcmc = NULL,
  first_yr = 1931,
  last_yr = NULL,
  n_proj = 5,
  n_iter = NULL,
  max.p = 5,
  max.d = 5,
  max.q = 5,
  arima = TRUE
)
}
\arguments{
\item{data}{a \code{list} of parameter values.}

\item{obj}{a \code{list} of parameter values.}

\item{mcmc}{a \code{list} of parameter values.}

\item{first_yr}{a \code{list} of inputs.}

\item{last_yr}{a \code{list} of inputs.}

\item{n_proj}{a \code{list} of inputs.}

\item{n_iter}{a \code{list} of inputs.}

\item{max.p}{Maximum value of p, or the maximum value of p (the AR order) to consider.}

\item{max.d}{Maximum value of d, or the maximum value of q (the MA order) to consider.}

\item{max.q}{Maximum value of q}

\item{arima}{default = TRUE, FALSE = "lognormal"}
}
\value{
a \code{list} of projected recruitment deviates and ARIMA specifications.
}
\description{
Project recruitment deviates
}
