% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_selectivity}
\alias{project_selectivity}
\title{Project recruitment deviates}
\usage{
project_selectivity(
  data,
  obj,
  mcmc = NULL,
  first_yr = 2000,
  last_yr = NULL,
  n_proj = 5,
  n_iter = NULL,
  arima = TRUE
)
}
\arguments{
\item{data}{a \code{list} of parameter values.}

\item{obj}{a \code{list} of parameter values.}

\item{mcmc}{a \code{list} of parameter values.}

\item{first_yr}{the first year.}

\item{last_yr}{the last year.}

\item{n_proj}{the number of projection years.}

\item{n_iter}{a \code{list} of inputs.}

\item{arima}{default = TRUE, FALSE = "lognormal"}
}
\value{
the negative log-likelihood (NLL) value.
}
\description{
Project recruitment deviates
}
